206221. 6.4.1.505. DbFunctions.Left(int?), DbFunctions.Right(int?); fix LikeCommon plan cache

+++ src/EntityFramework/Core/Objects/EntityFunctions.cs
@@ -560,7 +560,7 @@ namespace System.Data.Entity.Core.Objects
         [SuppressMessage("Microsoft.Usage", "CA1801:ReviewUnusedParameters", MessageId = "length")]
         [DbFunction("Edm", "Left")]
         [SuppressMessage("Microsoft.Naming", "CA1720:IdentifiersShouldNotContainTypeNames", MessageId = "string")]
-        public static String Left(String stringArgument, long? length)
+        public static String Left(String stringArgument, int? length)
         {
             throw new NotSupportedException(Strings.ELinq_DbFunctionDirectCall);
         }
@@ -580,7 +580,7 @@ namespace System.Data.Entity.Core.Objects
         [SuppressMessage("Microsoft.Usage", "CA1801:ReviewUnusedParameters", MessageId = "stringArgument")]
         [DbFunction("Edm", "Right")]
         [SuppressMessage("Microsoft.Naming", "CA1720:IdentifiersShouldNotContainTypeNames", MessageId = "string")]
-        public static String Right(String stringArgument, long? length)
+        public static String Right(String stringArgument, int? length)
         {
             throw new NotSupportedException(Strings.ELinq_DbFunctionDirectCall);
         }
